--- concordium-consensus/tools/block-state-dump/BlockStateDump/Util.hs
@@ -1,6 +1,0 @@
-module BlockStateDump.Util where
-
-import Control.Monad.IO.Class
-
-throwUserError :: (MonadIO m) => String -> m a
-throwUserError = liftIO . ioError . userError
